requirements.txt
flask
libtorrent
PyYAML